Bluetooth/movement conflict
When I add a movement function I start getting very strange things
happening, serial ports disappear, a println appeared to not copy
correctly to the Atmega, the code which previously worked (and was not
changed) no longer functions. Details in email to Joe.

TLDR- bluetoothMovement02 works fine. bluetoothMovement02FAIL does not.
It doesn’t even get to the code that was added because somehow the code
that used to work now doesn’t.